+++ packages/video_player/video_player/lib/video_player.dart
@@ -587,8 +587,15 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     _playerId =
         (await _videoPlayerPlatform.createWithOptions(creationOptions)) ??
         kUninitializedPlayerId;
-    _creatingCompleter!.complete(null);
-    final initializingCompleter = Completer<void>();
+    if (videoPlayerOptions != null) {
+      await _videoPlayerPlatform.setAllowScreenAutoLock(
+        _playerId,
+        videoPlayerOptions.allowScreenAutoLock,
+      );
+    }
+        videoPlayerOptions!.allowScreenAutoLock,
+      );
+    }
 
     // Apply the web-specific options
     if (kIsWeb && videoPlayerOptions?.webOptions != null) {