+++ packages/video_player/video_player_platform_interface/lib/video_player_platform_interface.dart
@@ -124,6 +124,13 @@ abstract class VideoPlayerPlatform extends PlatformInterface {
     );
   }
 
+  /// Sets whether the screen should be allowed to auto-lock during playback.
+  Future<void> setAllowScreenAutoLock(int playerId, bool allowScreenAutoLock) {
+    throw UnimplementedError(
+      'setAllowScreenAutoLock() has not been implemented.',
+    );
+  }
+
   /// Sets additional options on web.
   Future<void> setWebOptions(int playerId, VideoPlayerWebOptions options) {
     throw UnimplementedError('setWebOptions() has not been implemented.');
@@ -435,6 +442,7 @@ class VideoPlayerOptions {
   VideoPlayerOptions({
     this.mixWithOthers = false,
     this.allowBackgroundPlayback = false,
+    this.allowScreenAutoLock = false,
     this.webOptions,
   });
 
@@ -448,6 +456,12 @@ class VideoPlayerOptions {
   /// Note: This option will be silently ignored in the web platform (there is
   /// currently no way to implement this feature in this platform).
   final bool mixWithOthers;
+
+  /// Set this to true to allow the screen to auto-lock during video playback.
+  /// The default value is false, meaning the screen will stay awake during playback.
+  ///
+  /// This option is currently only supported on iOS and macOS.
+  final bool allowScreenAutoLock;
 
   /// Additional web controls
   final VideoPlayerWebOptions? webOptions;